created PlayerHandler
