+++ configs/config.py
@@ -5,7 +5,7 @@ VIDEOS = []
 FRAME_STEP = 5
 COUNT_PROCESSED_FRAMES = 0
 
-INDEX_OF_FRAME = 0
+INDEX_OF_FRAME = 5000
 INDEX_OF_VIDEO = 0
 INDEX_OF_All_FRAME = INDEX_OF_FRAME + (63600 * INDEX_OF_VIDEO)
 INDEX_OF_GPS = int(round(INDEX_OF_All_FRAME / 60, 0))
